added requirements, implemented fixes to permute added more labels for more datasets , matmul > mul

+++ project/model/train.py
@@ -23,7 +23,7 @@ def train(
 
             optimizer.zero_grad()
             yhat = net(data)
-            yhat.to(device).long()
+            yhat.to(device)
             loss = criterion(yhat, target)
             train_loss += loss.item()
